@@ -2,9 +2,7 @@ package epermit.controllers;
 
 import java.util.Comparator;
 import java.util.Optional;
-import java.util.stream.Collectors;
-import javax.validation.Valid;
-import org.springframework.security.access.prepost.PreAuthorize;
+import jakarta.validation.Valid;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -23,7 +21,6 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 @CrossOrigin(origins = "*")
 @RequestMapping("/authority_quotas")
-@PreAuthorize("hasRole('ADMIN')")
 public class QuotaController {
     private final AuthorityService service;
 
