@@ -2,18 +2,15 @@ package epermit.ledgerevents.quotacreated;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import javax.persistence.criteria.Predicate;
-import org.springframework.context.ApplicationEventPublisher;
+import jakarta.persistence.criteria.Predicate;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
-import epermit.appevents.QuotaCreated;
-import epermit.commons.Check;
+
+import epermit.commons.EpermitValidationException;
 import epermit.commons.ErrorCodes;
-import epermit.commons.GsonUtil;
 import epermit.entities.LedgerQuota;
+import epermit.ledgerevents.LedgerEventBase;
 import epermit.ledgerevents.LedgerEventHandler;
-import epermit.models.EPermitProperties;
 import epermit.repositories.LedgerQuotaRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
@@ -24,16 +21,15 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class QuotaCreatedLedgerEventHandler implements LedgerEventHandler {
     private final LedgerQuotaRepository quotaRepository;
-    private final EPermitProperties properties;
-    private final ApplicationEventPublisher eventPublisher;
 
     @Override
     @SneakyThrows
-    public void handle(Map<String, Object> claims) {
+    public <T extends LedgerEventBase> void handle(T claims) {
         log.info("QuotaCreatedEventHandler started with {}", claims);
-        QuotaCreatedLedgerEvent event = GsonUtil.fromMap(claims, QuotaCreatedLedgerEvent.class);
-        Long matched = quotaRepository.count(filterQuotas(event));
-        Check.assertTrue(matched == 0, ErrorCodes.INVALID_QUOTA_INTERVAL);
+        QuotaCreatedLedgerEvent event = (QuotaCreatedLedgerEvent) claims;
+        Boolean matched = quotaRepository.exists(filterQuotas(event));
+        if(matched)
+            throw new EpermitValidationException(ErrorCodes.INVALID_QUOTA_INTERVAL);
         LedgerQuota quota = new LedgerQuota();
         quota.setActive(true);
         quota.setEndNumber(event.getEndNumber());
@@ -44,15 +40,6 @@ public class QuotaCreatedLedgerEventHandler implements LedgerEventHandler {
         quota.setPermitIssuedFor(event.getEventProducer());
         log.info("QuotaCreatedEventHandler ended with {}", quota);
         quotaRepository.save(quota);
-        if (properties.getIssuerCode().equals(quota.getPermitIssuer())) {
-            QuotaCreated quotaCreated = new QuotaCreated();
-            quotaCreated.setEndNumber(event.getEndNumber());
-            quotaCreated.setStartNumber(event.getStartNumber());
-            quotaCreated.setPermitYear(event.getPermitYear());
-            quotaCreated.setPermitIssuedFor(event.getPermitIssuedFor());
-            quotaCreated.setPermitType(event.getPermitType());
-            eventPublisher.publishEvent(quotaCreated);
-        }
     }
 
     static Specification<LedgerQuota> filterQuotas(QuotaCreatedLedgerEvent event) {
@@ -62,11 +49,15 @@ public class QuotaCreatedLedgerEventHandler implements LedgerEventHandler {
             predicates.add(cb.equal(quota.get("permitIssuedFor"), event.getPermitIssuedFor()));
             predicates.add(cb.equal(quota.get("permitType"), event.getPermitType()));
             predicates.add(cb.equal(quota.get("permitYear"), event.getPermitYear()));
+            Predicate eventStartPredicate = cb.between(cb.literal(event.getStartNumber()),
+                    quota.get("startNumber"), quota.get("endNumber"));
+            Predicate eventEndPredicate = cb.between(cb.literal(event.getEndNumber()),
+                    quota.get("startNumber"), quota.get("endNumber"));
             Predicate startPredicate = cb.between(quota.get("startNumber"), event.getStartNumber(),
                     event.getEndNumber());
             Predicate endPredicate = cb.between(quota.get("endNumber"), event.getStartNumber(),
                     event.getEndNumber());
-            predicates.add(cb.or(startPredicate, endPredicate));
+            predicates.add(cb.or(eventStartPredicate, eventEndPredicate, startPredicate, endPredicate));
             return cb.and(predicates.toArray(new Predicate[predicates.size()]));
         };
         return spec;
