@@ -2,14 +2,15 @@ package epermit.controllers;
 
 import java.util.Map;
 import java.util.UUID;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import epermit.services.PrivateKeyService;
+import epermit.utils.PrivateKeyUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -17,13 +18,19 @@ import lombok.extern.slf4j.Slf4j;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/keys")
-@PreAuthorize("hasRole('ADMIN')")
 public class KeyController {
     private final PrivateKeyService keyService;
+    private final PrivateKeyUtil keyUtil;
+
+    @GetMapping()
+    public String export(){
+        return keyUtil.getKey().toJSONString();
+    }
 
     @PostMapping()
     public void create(@RequestBody Map<String, String> input) {
         log.info("Key create request. {}", input);
+        // Enable the key and disable all others  
         keyService.create(input.get("key_id"));
     }
 
