@@ -15,21 +15,14 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContext;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import epermit.entities.User;
 import epermit.models.dtos.PermitDto;
 import epermit.models.dtos.PermitListItem;
 import epermit.models.inputs.CreatePermitInput;
 import epermit.models.inputs.PermitUsedInput;
 import epermit.models.results.CreatePermitResult;
-import epermit.repositories.UserRepository;
 import epermit.services.PermitService;
 
 @ExtendWith(MockitoExtension.class)
@@ -48,8 +41,8 @@ public class PermitControllerTest {
             permits.add(dto);
         }
         Page<PermitListItem> pagedList = new PageImpl<>(permits);
-        when(permitService.getAll(any())).thenReturn(pagedList);
-        Page<PermitListItem> result = controller.getAll(Map.of());
+        when(permitService.getPage(any())).thenReturn(pagedList);
+        Page<PermitListItem> result = controller.getPage(Map.of());
         assertEquals(10, result.getTotalElements());
     }
 
@@ -82,10 +75,6 @@ public class PermitControllerTest {
 
     @Test
     void usedTest() {
-        Authentication authentication = Mockito.mock(Authentication.class);
-        SecurityContext securityContext = Mockito.mock(SecurityContext.class);
-        Mockito.when(securityContext.getAuthentication()).thenReturn(authentication);
-        SecurityContextHolder.setContext(securityContext);
         PermitUsedInput input = new PermitUsedInput();
         controller.setUsed("TR", input);
         verify(permitService, times(1)).permitUsed("TR", input);
