@@ -1,7 +1,5 @@
 package epermit.ledgerevents;
 
-import java.util.Map;
-
 public interface LedgerEventHandler {
-    void handle(Map<String, Object> event);
+    <T extends LedgerEventBase> void handle(T event);
 }
