@@ -14,7 +14,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 import epermit.appevents.LedgerEventCreated;
-import epermit.commons.Check;
+import epermit.commons.ApiErrorResponse;
+import epermit.commons.EpermitValidationException;
 import epermit.commons.ErrorCodes;
 import epermit.commons.GsonUtil;
 import epermit.entities.Authority;
@@ -45,9 +46,8 @@ public class LedgerEventUtil {
 
     public String getPreviousEventId(String consumer) {
         String previousEventId = "0";
-        Optional<LedgerEvent> lastEventR =
-                ledgerEventRepository.findTopByProducerAndConsumerOrderByCreatedAtDesc(
-                        properties.getIssuerCode(), consumer);
+        Optional<LedgerEvent> lastEventR = ledgerEventRepository.findTopByProducerAndConsumerOrderByCreatedAtDesc(
+                properties.getIssuerCode(), consumer);
         if (lastEventR.isPresent()) {
             previousEventId = lastEventR.get().getEventId();
         }
@@ -61,13 +61,12 @@ public class LedgerEventUtil {
         createdEvent.setSended(false);
         createdEventRepository.save(createdEvent);
         String proof = createProof(event);
-        handleEvent(GsonUtil.toMap(event), proof);
+        handleEvent(event, proof);
         publishAppEvent(createdEvent);
     }
 
     public LedgerEventCreated createAppEvent(CreatedEvent createdEvent) {
-        LedgerEvent ledgerEvent =
-                ledgerEventRepository.findOneByEventId(createdEvent.getEventId()).get();
+        LedgerEvent ledgerEvent = ledgerEventRepository.findOneByEventId(createdEvent.getEventId()).get();
         Authority authority = authorityRepository.findOneByCode(ledgerEvent.getConsumer());
         LedgerEventCreated appEvent = new LedgerEventCreated();
         appEvent.setId(UUID.randomUUID());
@@ -86,22 +85,24 @@ public class LedgerEventUtil {
     }
 
     @SneakyThrows
-    public void handleEvent(Map<String, Object> claims, String proof) {
-        log.info("Event handle started {}", claims);
-        LedgerEventBase e = GsonUtil.fromMap(claims, LedgerEventBase.class);
+    public <T extends LedgerEventBase> void handleEvent(T e, String proof) {
+        log.info("Event handle started {}", e);
         Boolean eventExist = ledgerEventRepository.existsByProducerAndConsumerAndEventId(
                 e.getEventProducer(), e.getEventConsumer(), e.getEventId());
-        Check.assertFalse(eventExist, ErrorCodes.EVENT_ALREADY_EXISTS);
+        if (eventExist)
+            throw new EpermitValidationException(ErrorCodes.EVENT_ALREADY_EXISTS);
+
         if (e.getPreviousEventId().equals("0")) {
             Boolean genesisEventExist = ledgerEventRepository
                     .existsByProducerAndConsumer(e.getEventProducer(), e.getEventConsumer());
-            Check.assertFalse(genesisEventExist, ErrorCodes.GENESIS_EVENT_ALREADY_EXISTS);
+            if (genesisEventExist)
+                throw new EpermitValidationException(ErrorCodes.GENESIS_EVENT_ALREADY_EXISTS);
             log.info("First event received");
         } else {
-            Boolean previousEventExist =
-                    ledgerEventRepository.existsByProducerAndConsumerAndEventId(
-                            e.getEventProducer(), e.getEventConsumer(), e.getPreviousEventId());
-            Check.assertTrue(previousEventExist, ErrorCodes.PREVIOUS_EVENT_NOTFOUND);
+            Boolean previousEventExist = ledgerEventRepository.existsByProducerAndConsumerAndEventId(
+                    e.getEventProducer(), e.getEventConsumer(), e.getPreviousEventId());
+            if (!previousEventExist)
+                throw new EpermitValidationException(ErrorCodes.PREVIOUS_EVENT_NOTFOUND);
         }
         LedgerEvent ledgerEvent = new LedgerEvent();
         ledgerEvent.setEventId(e.getEventId());
@@ -110,12 +111,11 @@ public class LedgerEventUtil {
         ledgerEvent.setEventTimestamp(e.getEventTimestamp());
         ledgerEvent.setEventType(e.getEventType());
         ledgerEvent.setPreviousEventId(e.getPreviousEventId());
-        ledgerEvent.setEventContent(GsonUtil.getGson().toJson(claims));
+        ledgerEvent.setEventContent(GsonUtil.getGson().toJson(e));
         ledgerEvent.setProof(proof);
         ledgerEventRepository.save(ledgerEvent);
-        LedgerEventHandler eventHandler =
-                eventHandlers.get(e.getEventType().toString() + "_EVENT_HANDLER");
-        eventHandler.handle(GsonUtil.toMap(claims));
+        LedgerEventHandler eventHandler = eventHandlers.get(e.getEventType().toString() + "_EVENT_HANDLER");
+        eventHandler.handle(e);
     }
 
     @SneakyThrows
@@ -142,8 +142,7 @@ public class LedgerEventUtil {
         String proof = authorization.substring(7);
         String[] proofArr = proof.split("\\.");
         String payloadJsonStr = GsonUtil.getGson().toJson(e);
-        String payloadBase64 =
-                Base64.getUrlEncoder().withoutPadding().encodeToString(payloadJsonStr.getBytes());
+        String payloadBase64 = Base64.getUrlEncoder().withoutPadding().encodeToString(payloadJsonStr.getBytes());
         String jws = proofArr[0] + "." + payloadBase64 + "." + proofArr[1];
         log.info("Constructed jws: {}", jws);
         log.info("Constructed jws length: {}", jws.length());
@@ -155,19 +154,21 @@ public class LedgerEventUtil {
     }
 
     @SneakyThrows
-    public Boolean sendEvent(LedgerEventCreated event) {
+    public ResponseEntity<?> sendEvent(LedgerEventCreated event) {
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
         headers.add("Authorization", "Bearer " + event.getProof());
         HttpEntity<Map<String, Object>> request = new HttpEntity<>(event.getContent(), headers);
-        ResponseEntity<?> result =
-                restTemplate.postForEntity(event.getUri(), request, Object.class);
-        if (result.getStatusCode() != HttpStatus.OK) {
-            MDC.put("epermitSendError", "SEND_EPERMIT_EVENT_ERROR");
+        ResponseEntity<?> result = restTemplate.postForEntity(event.getUri(), request, Object.class);
+        return result;
+        /*if (result.getStatusCode() != HttpStatus.OK) {
+            ApiErrorResponse error = (ApiErrorResponse)result.getBody();
+            if(error != null && error.getDetails().get("errorCode").equals("EVENT_ALREADY_EXISTS")){
+                
+            }
             log.error(GsonUtil.getGson().toJson(result.getBody()));
-            MDC.remove("epermitSendError");
             return false;
         }
-        return true;
+        return true;*/
     }
 }
