@@ -1,13 +1,12 @@
 package epermit.ledgerevents.keyrevoked;
 
 import java.util.List;
-import java.util.Map;
-import java.util.Optional;
 import org.springframework.stereotype.Service;
-import epermit.commons.Check;
+
+import epermit.commons.EpermitValidationException;
 import epermit.commons.ErrorCodes;
-import epermit.commons.GsonUtil;
 import epermit.entities.LedgerPublicKey;
+import epermit.ledgerevents.LedgerEventBase;
 import epermit.ledgerevents.LedgerEventHandler;
 import epermit.repositories.LedgerPublicKeyRepository;
 import lombok.RequiredArgsConstructor;
@@ -21,16 +20,15 @@ public class KeyRevokedLedgerEventHandler implements LedgerEventHandler {
     private final LedgerPublicKeyRepository publicKeyRepository;
 
     @SneakyThrows
-    public void handle(Map<String, Object> claims) {
+    public <T extends LedgerEventBase> void handle(T claims) {
         log.info("KeyRevokedLedgerEvent started with {}", claims);
-        KeyRevokedLedgerEvent e = GsonUtil.fromMap(claims, KeyRevokedLedgerEvent.class);
-        List<LedgerPublicKey> keys =
-                publicKeyRepository.findAllByAuthorityCodeAndRevokedFalse(e.getEventProducer());
-        Check.assertTrue(keys.size() > 1, ErrorCodes.INSUFFICIENT_KEY);
-        Optional<LedgerPublicKey> publicKeyR =
-                keys.stream().filter(x -> x.getKeyId().equals(e.getKeyId())).findFirst();
-        Check.assertTrue(publicKeyR.isPresent(), ErrorCodes.KEY_NOTFOUND);
-        LedgerPublicKey publicKey = publicKeyR.get();
+        KeyRevokedLedgerEvent e = (KeyRevokedLedgerEvent) claims;
+        List<LedgerPublicKey> keys = publicKeyRepository.findAllByAuthorityCodeAndRevokedFalse(e.getEventProducer());
+        if (keys.size() < 2)
+            throw new EpermitValidationException(ErrorCodes.INSUFFICIENT_KEY);
+        LedgerPublicKey publicKey = keys.stream().filter(x -> x.getKeyId().equals(e.getKeyId())).findFirst()
+                .orElseThrow(() -> new EpermitValidationException(ErrorCodes.KEY_NOTFOUND));
+
         publicKey.setRevoked(true);
         publicKey.setRevokedAt(e.getRevokedAt());
         log.info("KeyRevokedEventHandler ended with {}", publicKey);
