@@ -1,12 +1,11 @@
 package epermit.ledgerevents.permitrevoked;
 
-import java.util.Map;
-import java.util.Optional;
 import org.springframework.stereotype.Service;
-import epermit.commons.Check;
+
+import epermit.commons.EpermitValidationException;
 import epermit.commons.ErrorCodes;
-import epermit.commons.GsonUtil;
 import epermit.entities.LedgerPermit;
+import epermit.ledgerevents.LedgerEventBase;
 import epermit.ledgerevents.LedgerEventHandler;
 import epermit.repositories.LedgerPermitRepository;
 import lombok.RequiredArgsConstructor;
@@ -20,15 +19,18 @@ public class PermitRevokedLedgerEventHandler implements LedgerEventHandler {
     private final LedgerPermitRepository permitRepository;
 
     @SneakyThrows
-    public void handle(Map<String, Object> claims) {
+    public <T extends LedgerEventBase> void handle(T claims) {
         log.info("PermitRevokedEventHandler started with {}", claims);
-        PermitRevokedLedgerEvent event = GsonUtil.fromMap(claims, PermitRevokedLedgerEvent.class);
-        Optional<LedgerPermit> permitR = permitRepository.findOneByPermitId(event.getPermitId());
-        Check.assertTrue(permitR.isPresent(), ErrorCodes.PERMIT_NOTFOUND);
-        LedgerPermit permit = permitR.get();
-        Check.assertFalse(permit.isUsed(), ErrorCodes.PERMIT_USED);
-        Check.assertEquals(permit.getIssuer(), event.getEventProducer(), ErrorCodes.PERMIT_NOTFOUND);
-        Check.assertEquals(permit.getIssuedFor(), event.getEventConsumer(), ErrorCodes.PERMIT_NOTFOUND);
+        PermitRevokedLedgerEvent event = (PermitRevokedLedgerEvent) claims;
+        LedgerPermit permit = permitRepository.findOneByPermitId(event.getPermitId())
+                .orElseThrow(() -> new EpermitValidationException(ErrorCodes.PERMIT_NOTFOUND));
+        if (permit.isUsed())
+            throw new EpermitValidationException(ErrorCodes.PERMIT_NOTFOUND);
+        if (!permit.getIssuer().equals(event.getEventProducer()))
+            throw new EpermitValidationException(ErrorCodes.PERMIT_NOTFOUND);
+        if (!permit.getIssuedFor().equals(event.getEventConsumer()))
+            throw new EpermitValidationException(ErrorCodes.PERMIT_NOTFOUND);
+
         permitRepository.delete(permit);
     }
 }
