@@ -2,8 +2,6 @@ package epermit.utils;
 
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
-import java.io.FileNotFoundException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -11,8 +9,6 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.http.HttpStatus;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.itextpdf.barcodes.Barcode39;
 import com.itextpdf.barcodes.BarcodeQRCode;
 import com.itextpdf.kernel.pdf.PdfDocument;
@@ -20,9 +16,6 @@ import com.itextpdf.kernel.pdf.PdfWriter;
 import com.itextpdf.layout.Document;
 import com.itextpdf.layout.element.Image;
 import com.itextpdf.layout.element.Paragraph;
-import epermit.commons.ApiErrorResponse;
-import epermit.commons.EpermitValidationException;
-import epermit.commons.ErrorCodes;
 import epermit.models.EPermitProperties;
 import epermit.models.dtos.CreatePermitIdDto;
 import epermit.models.enums.PermitType;
@@ -48,7 +41,7 @@ public class PermitUtilTest {
         CreatePermitIdDto input = new CreatePermitIdDto();
         input.setIssuedFor("UA");
         input.setIssuer("TR");
-        input.setPermitType(PermitType.BILITERAL);
+        input.setPermitType(PermitType.BILATERAL);
         input.setPermitYear(2021);
         input.setSerialNumber(12);
         String permitId = util.getPermitId(input);
@@ -81,12 +74,12 @@ public class PermitUtilTest {
     void generateSerialNumberParameterizedTest(int endNumber) {
         /*
          * Authority authority = new Authority(); LedgerQuota quota = new LedgerQuota();
-         * quota.setPermitType(PermitType.BILITERAL); quota.setActive(true);
+         * quota.setPermitType(PermitType.BILATERAL); quota.setActive(true);
          * quota.setPermitYear(2021); quota.setEndNumber(endNumber);
          * authority.addIssuerQuota(quota);
          * when(issuedPermitRepository.findFirstByIssuedForAndRevokedTrue("UA")).thenReturn(Optional
          * .empty()); when(authorityRepository.findOneByCode("UA")).thenReturn(authority);
-         * Optional<Integer> result = util.generateSerialNumber("UA", 2021, PermitType.BILITERAL);
+         * Optional<Integer> result = util.generateSerialNumber("UA", 2021, PermitType.BILATERAL);
          * Assertions.assertEquals(2, result.get()); Assertions.assertEquals(endNumber == 3,
          * quota.isActive());
          */
